requirements.txt
fastapi
mongoengine
passlib[bcrypt]
pydantic[email,dotenv]
python-jose[cryptography]
python-multipart
uvicorn
